models/sbi_ariel_crossgen/requirements-vast.txt
astropy>=7,<8
bilby>=2.8,<3
configargparse>=1.7,<2
corner>=2.2,<3
glasflow>=0.4,<0.5
h5py>=3.10,<4
numpy>=1.26,<3
pandas>=2.2,<3
pyarrow>=17,<20
PyYAML>=6,<7
torchdiffeq>=0.2.5,<0.3
tqdm>=4.66,<5
wandb>=0.19,<0.21
threadpoolctl>=3.5,<4
